Move ScopedAdvancedTimer from AdvancedTimer.h to its own file.
